+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### 0.1.0 (Upcoming release)
+
 - Transportation problem definition
 - Linear programming solution of transportation problems 
 - Linear programming solution of shortest path problem
@@ -6,7 +7,7 @@
 - Linear programming solution of assignment problem
 - Linear programming solution of 2-players zero-sum game (Mostly copy-pasted from our JMcDM)
 - Default solver changed from GLPK to HiGHS
-  
+- p-median problem and its linear programming solution